+++ src/RcppExports.cpp
@@ -11,8 +11,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // mars_fit_cpp
-List mars_fit_cpp(const NumericMatrix& x_in, const NumericVector& y_in, int degree, int nk, double penalty, double thresh, int minspan_in, int endspan_in, int nprune, int pmethod, int trace, int nthreads);
-RcppExport SEXP _ares_mars_fit_cpp(SEXP x_inSEXP, SEXP y_inSEXP, SEXP degreeSEXP, SEXP nkSEXP, SEXP penaltySEXP, SEXP threshSEXP, SEXP minspan_inSEXP, SEXP endspan_inSEXP, SEXP npruneSEXP, SEXP pmethodSEXP, SEXP traceSEXP, SEXP nthreadsSEXP) {
+List mars_fit_cpp(const NumericMatrix& x_in, const NumericVector& y_in, int degree, int nk, double penalty, double thresh, int minspan_in, int endspan_in, int adjust_endspan, int auto_linpreds, int nprune, int pmethod, int trace, int nthreads);
+RcppExport SEXP _ares_mars_fit_cpp(SEXP x_inSEXP, SEXP y_inSEXP, SEXP degreeSEXP, SEXP nkSEXP, SEXP penaltySEXP, SEXP threshSEXP, SEXP minspan_inSEXP, SEXP endspan_inSEXP, SEXP adjust_endspanSEXP, SEXP auto_linpredsSEXP, SEXP npruneSEXP, SEXP pmethodSEXP, SEXP traceSEXP, SEXP nthreadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -24,11 +24,13 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type thresh(threshSEXP);
     Rcpp::traits::input_parameter< int >::type minspan_in(minspan_inSEXP);
     Rcpp::traits::input_parameter< int >::type endspan_in(endspan_inSEXP);
+    Rcpp::traits::input_parameter< int >::type adjust_endspan(adjust_endspanSEXP);
+    Rcpp::traits::input_parameter< int >::type auto_linpreds(auto_linpredsSEXP);
     Rcpp::traits::input_parameter< int >::type nprune(npruneSEXP);
     Rcpp::traits::input_parameter< int >::type pmethod(pmethodSEXP);
     Rcpp::traits::input_parameter< int >::type trace(traceSEXP);
     Rcpp::traits::input_parameter< int >::type nthreads(nthreadsSEXP);
-    rcpp_result_gen = Rcpp::wrap(mars_fit_cpp(x_in, y_in, degree, nk, penalty, thresh, minspan_in, endspan_in, nprune, pmethod, trace, nthreads));
+    rcpp_result_gen = Rcpp::wrap(mars_fit_cpp(x_in, y_in, degree, nk, penalty, thresh, minspan_in, endspan_in, adjust_endspan, auto_linpreds, nprune, pmethod, trace, nthreads));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -48,7 +50,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_ares_mars_fit_cpp", (DL_FUNC) &_ares_mars_fit_cpp, 12},
+    {"_ares_mars_fit_cpp", (DL_FUNC) &_ares_mars_fit_cpp, 14},
     {"_ares_mars_basis_cpp", (DL_FUNC) &_ares_mars_basis_cpp, 4},
     {NULL, NULL, 0}
 };